do not assume after/suggests modules have a relationship of after/suggests
* jhbuild/moduleset.py, tests/tests.py: do not assume after/suggests
modules have a relationship of after/suggests with their dependencies.
nautilus after tracker, nautilus-cd-burner depends on nautilus, tracker
depends on nautilus-cd-burner, the <after> would be evaluated first and
would transform the dependency of nautilus-cd-burner into an <after>,
ignoring it as it would cycle, causing nautilus-cd-burner to be built
before nautilus.  (closes: #563597)


svn path=/trunk/; revision=2566