change building state machine for an inter-phases dependency system
Use a simple dependency resolution system (no support for dependency
cycles) to build a list of phases required to build a module. This
allows more control over targets and puts a stop to the multiplication
of flags that were necessary to reach non-standard steps (such as
make distcheck).

The makeclean, makecheck, nobuild, makedist and makedistcheck options
can now all be implemented with a single build_targets option; support
exists to fill this new option according to the older flags.