[packagedb] don't ignore '' as version numbers
packagedb.check() allows None as version parameter to ignore the
version check; this patch changes the support so '' is not ignored
as well. This is useful for modules that were installed but got
their source directory removed, as jhbuild can't find a revision
number in that situation, and will pass the empty string.