Add support for using GConf schemas installed in the jhbuild prefix
Set GCONF_SCHEMA_INSTALL_SOURCE to point into the jhbuild prefix so
modules will install their schemas there (rather than failing to
install them into /etc).

Create a GConf source path file that tells GConf to use the data in
the jhbuild prefix (in addition to the data in the system prefix), and
point to it with GCONF_DEFAULT_SOURCE_PATH so modules will be read the
right data (assuming a new enough libgconf).

http://bugzilla.gnome.org/show_bug.cgi?id=586185