Fix testmodule
Trying to build a testmodule fails with a traceback ending with

 File "/usr/local/jhbuild/jhbuild/modtypes/testmodule.py", line 370, in parse_testmodule
    after=after, suggests=suggests, tested_pkgs=tested_pkgs)
TypeError: __init__() got an unexpected keyword argument 'suggests'

Remove the suggests argument to match the actual signature of __init__

Note this doesn't quite align with the documentation, which appears
to say that <suggests> is not permitted in a <testmodule> element.
This change has the effect of permitting <suggests>  but ignoring
the content.