[git] Use a new predicate to determine the need to stash
Introduce a predicate to test if the repository is dirty, optionally
including existing submodules.

This function is used to recognize uncommited changes to be stashed.
This should now work more reliable in the face of changes in submodules,
as the output of 'git diff --submodule' may have multiple lines, where
not every line starts with 'Submodule '.

To reliably exclude differences in submodules, a dependency to a two
year old git version is added.