[git] Switch branches previous to stash and pull
Now the refusal of the previously extracted function to switch dirty
branches will take effect. It is very unlikely that a user wants to
apply uncommited changes onto another branch by means of a jhbuild
update operation. Such an operation certainly deserves manual
inspection.

This will now update the new branch instead of the previous one.

The problem with not finding recently added remote branches is solved in
a previous commit.