Let sticky date move back and forth along master
Facilitate the bisection of a module, by making it possible to generate
a list of dependencies and move them back and forth in time after each
bisection step.
Previously it was only possible to move backwards in time.

Rename rewind_to_sticky_date to move_to_sticky_date to reflect this
change. There are no external users for this member.

https://bugzilla.gnome.org/show_bug.cgi?id=623930