[2.32,3.0] Move librsvgmm out of gnome-suites and to gnome
librsvg is in the desktop moduleset, not in the platform moduleset, so
librsvgmm should not be part of the official bindings. Since nothing
uses librsvgmm at the moment (as far as I know), we don't need to add it
to the desktop moduleset.

Also explicitly add atkmm and glibmm to the C++ bindings moduleset.