git: Fix switching to the currently configured dvcs-mirror
This takes the commands introduced by d1d8a2c and executes a single call
at the start of each update operation.  The repository will use the
mirror currently defined in jhbuildrc or the online location.

Previously it was not possible to switch from the online URL to a local
dvcs-mirror.

https://bugzilla.gnome.org/show_bug.cgi?id=647537