Add a "top_builddir" (defaults to $prefix/_jhbuild), change PackageDB to use it
For a future change to use "make install DESTDIR=", which in turn will
enable things like "jhbuild uninstall".

This patch adds an explicit concept in jhbuild of a directory for
build state called "top_builddir", distinct from the installation
tree.

We previously had something like this in the $prefix/share/jhbuild
directory, which is where it put the packagedb.xml file.

Move packagedb.xml in there, and migrate an old DB file.

https://bugzilla.gnome.org/show_bug.cgi?id=647231