autotools: Use make install DESTDIR=
Like most "package managers", use DESTDIR= to create a temporary
installation tree, from which we can generate a file manifest.

Add a <manifest> node to PackageDB which contains the contents of a
build.

Then implement "uninstall" using this.  A big advantage now is that we
don't need to do a build in order to do an uninstall, and uninstall is
independent of version.

https://bugzilla.gnome.org/show_bug.cgi?id=647231