Add new partial_build key, "sysdeps" command
Basically, when the partial_build key is set (which it is by default),
when we're gathering a module list (for say "jhbuild build", or just
"jhbuild list"), we look at the installed pkg-config files.  If there
is a local module which matches the id of the tarball, we skip it.

We should still be comparing versions.

But, this allows me to skip building e.g. cairo and nspr if I have
them on the system.

To allow developers to fill in using the system if available, add a
new "sysdeps" command.  This command parses the moduleset and finds
all .pc files that are needed by tarballs from the given moduleset,
and installs them.

https://bugzilla.gnome.org/show_bug.cgi?id=564373