Add post-installation trigger scripts
Several things in the GNOME stack have file-based caches, like
gdk-pixbuf's loaders.cache.  These need to be regenerated, and
historically that worked in jhbuild because these modules include bits
in their Makefile to run then relevant commands, if DESTDIR is not
set.

Package managers have post-installation scripts in packages, and OS
builders typically support this as well.

Jhbuild historically relied on the post-installation scripts in
modules, but in (see bug 647231) we switched to using DESTDIR for all
module types.  So we need to move closer to the package model here.

We call them "triggers" as they're run after any kind of build
operation (including "uninstall").

This patch adds a proof-of-concept trigger script for gdk-pixbuf
loaders.

https://bugzilla.gnome.org/show_bug.cgi?id=653842