moduleset: Only skip system dependencies if we haven't done a build before
If we're in partial_build mode, we need to keep building a module if
we've ever built it, even if the system dependency is now new enough.
An example is where one is using jhbuild and has built dbus, but then
one upgrades the distro, and a new enough system dbus comes in.

https://bugzilla.gnome.org/show_bug.cgi?id=654855