Delete no longer shipped files at install time
Many very difficult to debug build problems arise when a component at
some point in time ships a file (like a shared library), then later
switches to not shipping it, but the file remains in the jhbuild root.

This patch adjusts jhbuild to finally take advantage of the manifests
we have in the packagedb thanks to using make install DESTDIR.

I unified some work being done in 'jhbuild uninstall' with the install
process in a new 'fileutils.py'.

https://bugzilla.gnome.org/show_bug.cgi?id=654872