Enhance autogen.sh so only 1 way to install JHBuild (GNOME bug 654555)
If gnome-common and gnome-doc-utils are available, autogen.sh will
configure JHBuild to install via autotools.

If gnome-common and gnome-doc-utils are not available, autogen.sh will
configure JHBuild to install via a plain Makefile.

autogen.sh is used to configure JHBuild because the most common way to
obtain JHBuild is via git and a 'configure' should not be checked into
git.