packagedb: Lazily load cache
Only load the packagedb the first time someone calls get().  Refactor
the internals so that check() and installdate() both call get()
internally to avoid duplication.

This is a speed optimization (we instatiate the packagedb in cases
even if we're not going to read from it), as well as preparatory work
for locking.

https://bugzilla.gnome.org/show_bug.cgi?id=312910