build/buildone: Always run autogen, remove alwaysautogen config option
For what I'm fairly sure was originally a "speed" rationale, jhbuild
has up until now required a "-a" option to enable running autogen.sh.

However, one thing I am trying hard to eradicate from jhbuild is hard
to debug build failures.  And not running autogen.sh after we've done
a git pull, and then maybe relying on the module's AM_MAINTAINER_MODE
to rerun the autotools is a big source of very very hard to debug
failures.  There are actually some situations that 'make' cannot
reasonably detect (such as switching from recursive to nonrecursive
make).

Recently we added 'jhbuild make' which is more of the developer 'make
go fast' button one can use when actively hacking on a module.  It is
now explicitly defined to skip autogen.sh.

As far as the concern "but my builds are going to be slow!!!", I have
multiple answers to that:

0) We've already landed the partial build work, so you already have
   less to build than you did historically.
1) We are going to make up some of the speed by defaulting to parallel
   make, like we should have from day 0.  A patch already exists.
2) We can be more intelligent about detecting whether we need to run
   autogen.sh (this is a bit tricky but not impossible)
3) Longer term, jhbuild will be shipping binaries from gnome.org, so
   you will really only have to build what you modify.

https://bugzilla.gnome.org/show_bug.cgi?id=660844