nspr/nss: update to latest versions
The nss version we were shipping has a strange buildsystem that needs
separate configuration for each kernel version, so 3.x was causing
trouble.  The latest version fixes that.

The new nss needs a new nspr to go with (for PR_AF_INET6).

Our patches to nspr require running autoreconf, but nspr insists on
autoconf2.13, which would be annoying as an additional dependency, so we
bring back a patch to the generated 'configure' to avoid the need to run
autoreconf.

https://bugzilla.gnome.org/show_bug.cgi?id=669205