core-deps-3.4: fix symbol versioning in libxml2
The upstream release of libxml2 has a long-standing known bug with not
versioning its symbols correctly.  All the distos carry a patch for it
and this causes problems when system-installed libraries try to link
against our copy (which lacks the version information).

Start shipping the patch that everyone else is using.

https://bugzilla.gnome.org/show_bug.cgi?id=669437