use own rename() function that replaces existing files also on Windows
os.rename(src, dst) works differently on Unix and Windows.  On Unix, if
dst exists and is a file, it will be replaced silently if the user has
permission.  On Windows, if dst already exists, OSError will be raised.

All calls to os.rename() has been replaced with fileutils.rename(),
which tries to unlink the dst file, if exists, also on Windows.
Fixes Bug 670225.