distclean: Use git clean -d -f -x
Many forms of build failure are due to unclean source trees.  The
autotools are particularly susceptible to this.  For example,
previously when one did:

$ jhbuild buildone -afc glib

The jhbuild 'clean' operation ran *after* we ran configure, so
there can easily be stale libraries or binaries lying around
that were in the *old* Makefiles but not the *new* ones.

However, it would be dangerous to change the "clean" operation to be
"git clean -dfx" for developers; if they have unstaged/uncommitted
files lying around, they'll be deleted on a 'jhbuild make -afc'.

Therefore, I've added a new --distclean option.

This should *definitely* be used for tinderboxes.

https://bugzilla.gnome.org/show_bug.cgi?id=656081