bootstrap: Update libtool to 2.4.2 (GNOME bug 683408)
libtool 2.2.6a is getting fairly old (released 4 years ago), and
it does not work nicely with MinGW64. 2.4.2 is much newer, and does
not have this issue.

REMINDER: you really should be using the libtool version provided
by your distro, and avoid using jhbuild bootstrap as much as possible.
This reminder is valid for all tools built as part of jhbuild bootstrap.

However, for people who have/want to use jhbuild bootstrap, it's still
better to use a libtool version as good as possible.

https://bugzilla.gnome.org/show_bug.cgi?id=683408