gnome-suites-core-3.{6,8}.modules: make gnome-shell hard require gdm
Though gdm is not a build pre-requisite of gnome-shell, if we don't
build gdm at all, the resulting tree will not function on
pre-GNOME-3.6 systems, since GNOME Shell has a hard requirement
on the existence of GDM library introspection.

The result will work even when the built GDM is not installed on
the systems, since the libgdm code is tolerant of such a mismatch.

https://bugzilla.gnome.org/show_bug.cgi?id=685351