uninstall: support automatically created modules
"jhbuild buildone" creates automatic modules for arguments that are not
declared modules in the moduleset.

Once installed, these can not be uninstalled because "jhbuild uninstall"
demands modules to in the moduleset file, even when manifest files are
available for the automatically created ones.

Fix uninstall to support on-the-fly creation of modules. This works just
like with moduleset modules.

https://bugzilla.gnome.org/show_bug.cgi?id=687601