Fix performance bottleneck in jhbuild tinderbox
The HTML logging for jhbuild tinderbox contained code that would read
1024 bytes of build stdout and then sleep for 0.1 seconds.  This
effectively rate-limited builds to producing 10k of output per second,
which is easily overrun by chatty builds like WebKit on fast systems.

This commit changes the buffer size to 10000 and eliminates the sleep
entirely.  It also reworks the handling of when an output line is
printed, always waiting until a newline is seen, except at the very end
of builds.

This prevents this sort of output:

    checking for sys/types.h...
    yes

that was often seen (on separate lines) in the logfiles before.

https://bugzilla.gnome.org/show_bug.cgi?id=689194