Fix handling of trailing slashes on command-line modules names
It is currently possible to delete the whole checkoutroot by adding
a trailing slash to a module name passed on the command line:

• Run `jhbuild buildone gnome-desktop/`
• Watch jhbuild error out as it is unable to find the repo
• Select 'Wipe directory and start over' (just kidding, *don't*
  select it!)
• Watch jhbuild delete your whole checkout root.

This happens because the module name is passed to os.path.basename()
which will return an empty string on trailing slashes.

JHBuild will then try to operate on the base checkout root, failing to
find any proper repo and promptly deleting all its content if told to
do so.

This patch does two things to prevent this error to occur:

• Strip all the trailing slashes from module names in
  Branch.get_module_basename()
• Refuse to accept an empty string for the module basename in
  Branch.get_checkoutdir()

It also tries to strip trailing slashes from module names in a couple of
other places, trying to keep stuff working even when being passed one.

https://bugzilla.gnome.org/show_bug.cgi?id=694466