3.10: mutter-wayland depends on the XWayland server
mutter-wayland won't start in wayland mode unless it can spawn
an Xorg binary compiled with wayland support. And the X server
needs at least one driver.
Because the only driver that was updated after the xserver rebase
is the intel one, that is included in the moduleset. Sorry people
if you run any other card.