install: remove indentation to comment, so it doesn't get printed
In Makefiles, if a comment starts with the tab, it will get printed to
the output like if it was an echo statement. Fix this, to avoid confusing
the people at 'make install' (as this is a comment to clarify the
subsequent lines of code, not a comment for the jhbuild user).

https://bugzilla.gnome.org/show_bug.cgi?id=706518