autogen.sh: drop bashism, use POSIX sh
We had a complicated loop in autogen.sh for parsing commandline options.
It allowed setting arbitrary variables in the script, despite the fact
that only two variable were ever documented as being supported (or
probably ever used).  The ability to set random shell variables was
pretty questionable to begin with.

Remove the bash-specific substring operations and replace the loop with
a more common switch-based approach which should work with any POSIX
shell (tested with dash).  Change to #!/bin/sh.

This patch has the side effect that the user can no longer use a
commandline argument to modify the IFS of the shell running autogen.sh,
but that seems like a good thing to me.

https://bugzilla.gnome.org/show_bug.cgi?id=720068