moduleset: teach jhbuild how to build itself
jhbuild is perfectly capable of building itself, so add it to the
moduleset.

It was always possible to use 'jhbuild buildone jhbuild' to do it
before, but having it in the moduleset lets us specify the correct
dependencies and allows us to build jhbuild from other commands, such as
tinderbox.

This will allow people to bootstrap their setup using a 'simple install'
jhbuild (that doesn't need yelp or gnome-common) followed by a normal
full install, with dependencies.  This is helpful for systems that don't
have yelp-tools or gnome-common.

https://bugzilla.gnome.org/show_bug.cgi?id=721323