ibus: update to 1.5.5 and add a patch
Recent changes to 'dconf update' combined with ibus's strange use of
dconf mean that ibus fails to install if the system is missing an
/etc/dconf/db directory.

Update to the latest tarball release from upstream and cherry-pick a
patch from their git in order to work around the issue.

We can also take this chance to drop our -Wno-error because this problem
was solved upstream in the latest tarball.

https://bugzilla.gnome.org/show_bug.cgi?id=722978