autotools: Add configure to the error phases for the install phase
Changes in other modules may cause the install phase to fail since the
configure phase has not been run to pick up the changes.

Add the configure phase to the list of error phases for the install
phase, so that it can be inreactively choosen at the jhbuild prompt for
failing modules or, if the 'trycheckout' configuration option is
enabled, configure can be run to try to automatically recover from a
failing module.

https://bugzilla.gnome.org/show_bug.cgi?id=723083