apps: clean up spice, drop celt, use opus
Update to the latest release of spice-gtk, which (after much
anticipation) supports opus instead of the long-unmaintained celt051.

Drop spice-protocol as a dependency of spice-gtk since it is now rolled
into the tarball upstream (via gitsubmodules).

Drop spice-server as well because nothing in the entire moduleset
actually uses it.

Add a sysdep for opus.

https://bugzilla.gnome.org/show_bug.cgi?id=724116