setup_env(): remove partial_build conditions
Don't consult the partial_build variable when deciding if or if not to
include the default paths (/usr/local, /usr) in the aclocal and
pkg-config paths.

This is a leftover from the days that it was actually reasonably
possible to bootstrap yourself into a complete jhbuild world without
depending on anything on the system at all.