autotools: remove '--' GStreamer hack
This strange hack looks like it might have gotten introduced by accident
as part of e9084288cdf1e520c409ca8fb377480676e46782 which is otherwise a
completely unrelated commit.  I have no idea why this might have been
needed, but it seems very likely that if it was, it is no longer.