refactor sysdep uninstalled list handling a bit
Instead of collecting uninstall sysdeps into two lists (one for pkgconfig and
one for filenames) use just one list along with a 'dep_type' field.

This lets us send the list into the backend modules without converting C
includes and path programs into absolute filenames.

Right now we have a convenience function in the backend that implements the
same logic as before, but this can be improved in the future.

This will also allow for easier addition of new types of sysdeps.

https://bugzilla.gnome.org/show_bug.cgi?id=691394