sysdeps: support for 'xml' and 'python2' sysdeps
Add support for two new types of system dependencies.

An 'xml' dependency names a uri that must exist in the local xml catalog.  We
discover the path of the xml catalog using the same logic that gtk-doc uses
(which is appropriate since gtk-doc is the only program requiring xml depends
at the moment).

A 'python2' dependency names a python module that should be available for use
with 'import' in python 2.

There is not currently support in any of the 'sysdep --install' backends for
installing these sysdeps if they are missing, but at least now we can report
them as missing instead of just waiting for the build to fail.

https://bugzilla.gnome.org/show_bug.cgi?id=691394