3.16: fill in a lot of base system depends
Add sysdep packages for gettext-bin, xmlcatalog, pkg-config, python2,
cc, libxml2-python, docbook-xsl, xmllint and xsltproc.

The docbook-xsl and libxml2-python packages take advantage of the
newly-added support for xml and python package sysdeps.

This brings us quite close to being able to fully list off the packages
required for installation into a base Debian system in order to build
glib.

automake is not explicitly listed here because it gets pulled in by
intltool anyway (which pulls in make and a bunch of other nice stuff).
We may want to make this more explicit in the future, but see below.

We semi-arbitrarily list libtool, gettext-bin and cc on gtk-doc because
this is the first package that requires these things and it gets pulled
in by almost anything else that anyone would want to build (by way of
glib).

We do not list 'git', even though it is required to build almost
anything.  We also do not list 'pkg-config' or 'xmlcatalog' on the
packages that require those to be around for detecting the existence of
other packages.  The best approach here would probably be to
automatically add dependencies on these programs in case we find a
package that will require them (eg: 'git' is required for any package
downloaded from git).  We can also add a proper dependency on 'automake'
in the same way.

'pkg-config' and 'xmlcatalog' are listed as dependencies of jhbuild
itself because if you use jhbuild to build itself then you really should
have these programs available for jhbuild itself to use after it is
installed.

Some of these dependencies seem pretty theoretical (like checking if
python is installed from a program written in python) but the
completeness will allow us to easily generate a full list of all
software that is required for using jhbuild to build a particular set of
modules.  This will be useful for populating build chroots, among other
things.

https://bugzilla.gnome.org/show_bug.cgi?id=741538