Fix security warning when using sysdeps --install
Attempting to run 'jhbuild sysdeps --install' on Fedora 21:

"This software is not from a trusted source. Do not install this
software unless you are sure it is safe to do so."

The security warning is due to misuse of PackageKit's API, which changed
a couple of years ago. Thanks to Richard Hughes for help debugging.

https://bugzilla.gnome.org/show_bug.cgi?id=741774