config: add condition flags for 'darwin'
Detect 'darwin' in sys.platform and assume that it means that we're on
Mac OS X.

Define new condition flags 'darwin', 'macos', and 'quartz' for that.

We already refer to the 'macos' condition in the bootstrap moduleset but
so far it had never been defined.

The intention is that 'macos' will be used for expectations about the
software available in the system (and what needs to be bootstrapped) and
that 'quartz' will refer to the window system.  It would theoretically
make sense to disable 'quartz' and enable 'x11' if one wanted to use the
X server on macos, for example.

We additionally add a 'gnu-elf' condition for systems that have
GNU-style binutils (which accept -Bsymbolic, for example).