bootstrap: remove vala-bootstrap from meta-module
Remove vala-bootstrap from the list of modules build during 'jhbuild
boostrap' since it will fail (due to not-yet-installed GLib).

After GLib is installed, the user will need to type 'jhbuild bootstrap
vala-bootstrap' in order to get things working properly.

This is not a great solution but it is better than the bootstrap
aborting half way through because of a module that has no chance of
building.