sysid: handle condition sets
Move the determination of the default conditions for the system to a
better home in the sysid module.  Add conditions for the ID and ID_LIKE
that we find in the os-release file.  That means, for example, that an
Ubuntu system will end up with "debian" and "ubuntu" conditions set by
default.

https://bugzilla.gnome.org/show_bug.cgi?id=742267