3.16: turn polkit into a system module
polkit depends on spidermonkey, which is API unstable.  There is no
version of spidermoney which polkit can use which is available in all of
the distributions that we care about supporting, which means that we
have trouble with sysdeps.

Fortunately, every distribution that we care about already has polkit
for itself.  polkit itself is very inactive these days -- the highest
dependency that any GNOME module has on it is for a version that was
released in 2011.  Additionally, we are not actually interested in
building polkit itself, but really only need the libraries.  It would be
better if we used the system ones.

A survey of the configure.ac files of components that declare a
dependency on polkit shows:

 - udisks: depends on polkit-gobject and polkit-agent 0.102
 - gnome-control-center, gnome-settings-daemon: polkit-gobject 0.103
 - colord: polkit-gobject 0.103
 - accountsservice: polkit-gobject
 - NetworkManager: polkit-agent 0.97
 - ModemManager polkit-gobject 0.97
 - PackageKit polkit-gobject 0.98
 - gnome-shell: polkit-agent 0.100
 - pulseaudio, gnome-session: no longer actually depend on it

So let's get rid of the polkit module and add two new system modules:
one for each .pc file.  We must have at least polkit-agent 0.102 and
polkit-gobject 0.103.

https://bugzilla.gnome.org/show_bug.cgi?id=742269