packagedb: write only one manifest
Because of the fact that the 'manifest' attribute on PackageEntry items
is a property and because of the fact that we check it each time we
convert a node to XML, we end up reading every manifest file every time
we go to update the packagedb (ie: every time we install a package).

Even worse: because self.manifest is now populated, we also write every
single manifest file each time we update the packagedb.

Avoid the problem by making the write explicit for the module being
added, and no other.

https://bugzilla.gnome.org/show_bug.cgi?id=742298