MakeModule: add 'make' helper
Instead of repeating the logic in a dozen places, add a helper function
that collects arguments, builds the 'make' invocation and executes it.

Use this from the autotools and cmake modtypes.

There is an inconsistency about whether makeargs are passed to 'make
install' between autotools (they aren't) and cmake (they are).  This
patch preserves that inconsistency.

https://bugzilla.gnome.org/show_bug.cgi?id=742292