MakeModule: add needs_gmake flag
Add a needs_gmake flag for MakeModule.

If this is set to True (which is the default for autotools) and the
'gmake' condition is set then the 'gmake' command will be used instead
of 'make'.

This helps with systems where the system 'make' is not GNU make.

This patch effectively establishes an assumption that all autotools
packages will require gmake.  In the unlikely event that this proves to
cause trouble for someone, we could easily add a needs-gmake attribute
to the XML to allow people to set it back to 'false'.

Since the make command might now be 'gmake', make sure we take the
correct one for the creation of virtual sysdeps.

https://bugzilla.gnome.org/show_bug.cgi?id=742292