core: rewrite the dependency solver, again
Although this was already done a couple of years ago (bug 669554), there
is room for more simplicity in the algorithm, with only small and
desirable functionality changes.

The new algorithm offers much more consistent handling of 'after'
depends (if they are enabled): it simply installs them after the module
in question, as if the user had requested that.  It also removes
warnings about cyclic dependencies when 'A' depends on 'B' and 'B' is
after 'A'.  This is a perfectly reasonable situation.

The change in handling of 'after' depends will produce some surprising
results when we enable 'after' by default.  For example, 'dconf' is an
'after' of 'glib', but depends on 'gtk+' (for the editor) so building
'glib' will pull in 'gtk+'.  That's perfectly logical -- we'll just need
to figure out a way to sort it all out.

The new algorithm will also abort in case a dependency cycle is detected
while attempting to satisfy the hard depends of a module requested for
building which is the only reasonable thing to do, in my opinion.

https://bugzilla.gnome.org/show_bug.cgi?id=693290