core: stop ignoring 'after' modules
The --build-optional-modules flag on the 'build' command was the only
way to get jhbuild to actually consider 'after' modules.  Get rid of it
and turn that functionality on by default.

Add a new config flag to turn it off.  That makes it the same as
'suggests' now (which has a similar config flag).

https://bugzilla.gnome.org/show_bug.cgi?id=693290