systeminstall: speed up pkg-config
The code for building the list of all installed packages is very slow,
particularly when lots of packages are installed.

This "slow and steady" approach was implemented in bug 683832 because of
the potential for broken pkg-config files to break the old approach.

Broken pkg-config files are not normally installed, so we may as well
try to do things the fast way first, falling back on the other approach
if it fails.

This improves 'jhbuild list glib' from ~5.6 to ~0.4 seconds on my fairly
typical system (with 482 .pc files installed).

https://bugzilla.gnome.org/show_bug.cgi?id=740866