packagedb: store relative paths in manifest
This change does not actually touch any packagedb code, but rather
changes its users to assume that the 'manifest' property is populated
with relative pathnames.

All readers pass the list through fileutils.filter_files_by_prefix() so
modify that to take either paths relative to the given prefix or
absolute paths (which will be filtered to limit them to the prefix).

The code in process_install() is the only writer, so simply modify it to
use relative paths.

Although much of the software that jhbuild builds does not like to have
its install prefix moved (due to hardcoding of paths), this change means
that jhbuild itself theoretically has no problem with it.

It also makes the process of generating tar files using the manifest
slightly cleaner.

https://bugzilla.gnome.org/show_bug.cgi?id=743127