config: add disable_Werror config option
Add a disable_Werror configuration option (default True) which causes
--disable-Werror to be passed to each automake build.

Support for this switch was recently added to AX_COMPILER_FLAGS as a way
to unconditionally disable all fatal warnings.  By default, -Werror is
enabled for builds done out of git (which is what jhbuild spends most of
its time doing).

The idea is to prevent -Werror from ruining the day for people who
simply want to use jhbuild as a mechanism to build the latest version of
GNOME, while allowing advanced users to re-enable the feature (and also
maintaining -Werror for those who do manual git checkouts).

In the case that the module doesn't know about this switch, it will be
safely ignored.