tarball: delay check for writable tarballdir
It's possible to imagine jhbuild used in a situation where the tarball
directory may not be writable by the user who is running it, including
in the case where the tarballdir is a read-only bind mount.

For that reason, skip the check on startup for tarballdir being
read-only and delay it to the point where we would actually try to
download the file.

https://bugzilla.gnome.org/show_bug.cgi?id=746327