Delete GNU Texinfo dir file and handle it with a trigger
Modules with GNU Texinfo documentation creates new dir files during
'make install', which causes the installed dir file to be replaced with
the newly generated dir file with only one entry. To make GNU Texinfo
usable in JHBuild, we delete the dir file in DESTDIR and handle it
with a new trigger.

It is possible to have unavailable entries in the dir file because
triggers are not run when uninstalling modules. It will be fixed if
'jhbuild postinst' is run or other modules with Texinfo documentation
are successfully built.

https://bugzilla.gnome.org/show_bug.cgi?id=750484