sysdeps: Support alternative dependencies
Some system dependencies, such as LLVM, cannot be easily detected with
single c_include or path because different distributions can install the
same things in different places with different names.

To resolve this problem, we add a new type of tag <altdep>, which can be
used as a child node of <dep> in <systemdependencies> to specify the
list of alternative c_include and path. The syntax of <altdep> and <dep>
are the same, so we can reuse code used to process <dep>.

https://bugzilla.gnome.org/show_bug.cgi?id=754041