3.20: Make cairo a sysdep
I'm a bit hesitant because GTK+ requires a cairo that is only one year
old, and this dependency is bound to increase in the future. But we have
to break the circular dependency gobject-introspection -> cairo ->
librsvg -> gdk-pixbuf -> gobject-introspection or new users will never
be able to build anything, and the best way I see to do that is to take
cairo out of the equation. If this doesn't work we might need to explore
building two separate cairos, one as a bootstrap without SVG support.