install-check: Drop as no longer useful and potentially dangerous
`install-check` is totally pointless when installing to a DESTDIR and
copying the files into place, which nowadays is JHBuild's normal method
of operation.

It's also potentially problematic when mixed with .pyc files in the case
that we don't do a DESTDIR install (for some reason).  The .pyc file can
be updated (due to the fact that it contains a timestamp) and will
therefore be installed, but the original .py file wouldn't be because
it's the same.  This means that the timestamp in the .pyc file would
mismatch the installed .py file (which would either be equal to the
timestamp that it was checked out of git, or that it was generated, if
that is the case).

https://bugzilla.gnome.org/show_bug.cgi?id=721167