3.20: Move speech-dispatcher back to core-deps
This .pc file doesn't exist in Fedora. Basically the dleyna-renderer
situation all over again, except this time the .pc file does not seem to
be removed intentionally... in fact, I have not the faintest clue why it
is missing.

It really should be in sysdeps, but move it to core-deps until we can
figure out how to make it work in sysdeps without breaking Fedora.