autotools: add a "bootstrap" attribute to skip default dependencies
Packages like autoconf/automake in the bootstrap moduleset use
the autotools module type which defines default dependencies
and leads to dependency cycles during bootstrap.

This adds a bootstrap attribute which removes the default dependencies
and allows the packages to control all dependencies.

Initially the plan was to re-use the skip-autogen attribute and skip
the default dependencies if "true", but skip-autogen also prevents
configure from running which is still needed for those packages.