sysdeps-3.20: Add libusb 1.0.9 to keep compatibility with FreeBSD
FreeBSD seems to use a libusb implementation that is different from
libusb found on other operating systems, so the latest version may
be unavailable. FreeBSD 10.x (release and stable snapshot) only has
libusb 1.0.9 and FreeBSD 11.0 (development snapshot) has libusb 1.0.13.
To keep JHBuild working on FreeBSD, a system module called libusb1-1.0.9
is added to lower the version requirement.

Move usbredir and gnome-boxes to libusb1-1.0.9 because this is the
version requirement they set in their configure.ac.

Make libusb dependency of colord optional because it is only used
when libgusb support is enabled.