3.22: gtk+ should suggest desktop-file-utils
Many GTK+ apps depend on desktop-file-utils, but we don't know how many
and none currently have the dependency specified. Since it's a sysdep,
we can just add it to gtk+ and be done with it. This fixes the problem
for all except brave users who disable suggests; such users are expected
to be able to handle breakage caused by that.

https://bugzilla.gnome.org/show_bug.cgi?id=767372