3.24: Remove WebKit1 and dependencies
I'm removing Empathy since the latest tarball still depends on
clutter-gst2 and it doesn't pass distcheck for reasons I don't
understand when trying to release it.

Since Empathy is the last module outside unsupported world that depends
on WebKit1, that means we can remove it at long last.

Note this means Geary gets removed as fallout. Geary should really be
added to gnome-apps, it doesn't deserve to be relegated to -world, but
that cannot happen until it moves to WebKit2.