Fix 'jhbuild sysdeps' when ninja is not installed
This regressed in 73b433d89cd4c2496e9d44659a486d4ea0e2a20c, which is sad
because it means nobody has bothered to report that our Build GNOME
guide has been broken for the past two months.

We cannot check for ninja binary in __init__ because it gets called from
sysdeps, so we end up failing because the thing we're trying to install
is not already installed. Check before each point of use instead.