systeminstall: enhance slowness warning for apt backend
It's always been way slower than PackageKit (at least the Fedora
backend), and I just made it even slower because now it checks for
almost all include files twice. So let's make the warning text more
clear here, to encourage users not to worry that it has hung. I was
suspicious myself, when jhbuild was doing just fine and was just slow.